Code review feedback - Using VersionFolderPathResolver for manifest file name - Moved response file parsing to seperate class
